# src/swagger/swagger.json
{
	"openapi": "3.0.3",
	"info": {
		"title": "Fortress API",
		"version": "1.0.0",
		"description": "Fortress is a social media platform that allows men to connect, share experiences, and support each other in a safe and empowering environment. The platform is designed to foster positive masculinity, mental health awareness, and community engagement through various features such as posts, comments, and user profiles. This is achieved through anonimity, robust moderation, and a focus on constructive interactions."
	},
	"paths": {
...
		"/like/user/{userId}": {
			"get": {
				"tags": ["Like"],
				"summary": "Get likes by user",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "page",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"schema": {
							"type": "integer",
							"default": 20
						}
					}
				],
				"responses": {
					"200": {
						"description": "User likes",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LikesListResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/story/user/{userId}": {
			"get": {
				"tags": ["Story"],
				"summary": "Get user stories",
				"parameters": [
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "User stories",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Story"
									}
								}
							}
						}
					}
				}
			},
			"/story/{id}": {
				"get": {
					"tags": ["Story"],
					"summary": "Get story by id",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Story",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/StoryResponse"
									}
								}
							}
						},
						"404": {
							"description": "Not found",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"410": {
							"description": "Expired",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				},
				"post": {
					"tags": ["Story"],
					"summary": "View a story",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Viewed",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/StoryViewResponse"
									}
								}
							}
						},
						"404": {
							"description": "Not found",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"401": {
							"description": "Unauthorized",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				},
				"delete": {
					"tags": ["Story"],
					"summary": "Delete story",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Deleted",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/GenericMessageResponse"
									}
								}
							}
						},
						"403": {
							"description": "Forbidden",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"401": {
							"description": "Unauthorized",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				}
			},
			"/story/{id}/viewers": {
				"get": {
					"tags": ["Story"],
					"summary": "Get story viewers",
					"parameters": [
						{
							"name": "id",
							"in": "path",
							"required": true,
							"schema": {
								"type": "string"
							}
						}
					],
					"security": [
						{
							"bearerAuth": []
						}
					],
					"responses": {
						"200": {
							"description": "Viewers list",
							"content": {
								"application/json": {
									"schema": {
										"type": "object",
										"properties": {
											"viewers": {
												"type": "array",
												"items": {
													"$ref": "#/components/schemas/User"
												}
											},
											"totalViews": {
												"type": "integer"
											}
										}
									}
								}
							}
						},
						"403": {
							"description": "Forbidden",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						},
						"401": {
							"description": "Unauthorized",
							"content": {
								"application/json": {
									"schema": {
										"$ref": "#/components/schemas/ErrorResponse"
									}
								}
							}
						}
					}
				}
			}
		},
		"/user/me": {
			"get": {
				"tags": ["User"],
				"summary": "Get current user",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Current user",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["User"],
				"summary": "Update current user's profile",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "User updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"400": {
						"description": "Validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidationErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/user/{id}": {
			"get": {
				"tags": ["User"],
				"summary": "Get user by id",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "User found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"403": {
						"description": "Private profile",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/user/search/{query}": {
			"get": {
				"tags": ["User"],
				"summary": "Search users",
				"parameters": [
					{
						"name": "query",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Search results",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/User"
									}
								}
							}
						}
					}
				}
			}
		},
		"/user/suggestions/{id}": {
			"get": {
				"tags": ["User"],
				"summary": "Get user suggestions",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Suggestions",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/User"
									}
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/user/private": {
			"post": {
				"tags": ["User"],
				"summary": "Toggle private account",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Toggled privacy",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserResponse"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
...
			"User": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"profile": {
						"$ref": "#/components/schemas/UserProfile"
					},
					"counts": {
						"$ref": "#/components/schemas/UserCounts"
					},
					"isVerified": {
						"type": "boolean"
					},
					"isPrivate": {
						"type": "boolean"
					},
					"lastLoginAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"UserProfile": {
				"type": "object",
				"properties": {
					"fullName": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": ["male", "female", "other", "prefer_not_to_say"]
					}
				}
			},
			"UserCounts": {
				"type": "object",
				"properties": {
					"posts": {
						"type": "integer"
					},
					"followers": {
						"type": "integer"
					},
					"following": {
						"type": "integer"
					}
				}
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
			"ValidationErrorResponse": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"error": {
						"type": "string"
					},
					"okay": {
						"type": "boolean"
					}
				}
			},
...
			"GenericMessageResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"type": "string"
					}
				}
			},
...
			"Like": {
				"type": "object",
				"properties": {
					"_id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"targetType": {
						"type": "string"
					},
					"targetId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"LikesListResponse": {
				"type": "object",
				"properties": {
					"likes": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Like"
						}
					},
					"pagination": {
						"type": "object",
						"properties": {
							"page": {
								"type": "integer"
							},
							"limit": {
								"type": "integer"
							},
							"total": {
								"type": "integer"
							},
							"totalPages": {
								"type": "integer"
							}
						}
					}
				}
			},
...
			"UpdateUserProfileDto": {
				"type": "object",
				"properties": {
					"fullName": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"avatar": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"gender": {
						"type": "string",
						"enum": ["male", "female", "other", "prefer_not_to_say"]
					}
				}
			},
			"UpdateUserDto": {
				"type": "object",
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"isPrivate": {
						"type": "boolean"
					},
					"profile": {
						"$ref": "#/components/schemas/UpdateUserProfileDto"
					}
				}
			},
			"UserResponse": {
				"type": "object",
				"properties": {
					"okay": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "integer"
					},
					"message": {
						"$ref": "#/components/schemas/User"
					}
				}
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
...
	"security": [{ "bearerAuth": [] }]
}